Add Q&A app with answer-graph visualisation, notebooks, and docs
- Streamlit chat: natural-language Q&A over the Neo4j graph, with an
  interactive subgraph rendered alongside each answer (src/viz.py via pyvis)
- Fix schema/loader type mismatches: caseNumber and Section division/group/
  classCode now load as INTEGER and match the schema given to the LLM
- Add data-pipeline notebooks (data/Cleaning/analysis/knowledge_graph) + data
- Add proposal (docs/), comprehensive README, and .env.example
- Remove dead main.py; tidy .gitignore and package metadata

+++ src/schema.py
@@ -4,8 +4,8 @@ Node properties:
         caseInitiationDate: DATE, caseLastDecisionDate: DATE,
         displayName: STRING}
 - Company {name: STRING, displayName: STRING}
-- Section {name: STRING, sectionCode: STRING, division: FLOAT,
-           group: FLOAT, classCode: FLOAT, classDescription: STRING,
+- Section {name: STRING, sectionCode: STRING, division: INTEGER,
+           group: INTEGER, classCode: INTEGER, classDescription: STRING,
            displayName: STRING}
 - LegalBasis {name: STRING, displayName: STRING}
